Added new migrations and implemented user registration

+++ api/Controllers/StockController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.DTOs;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using Microsoft.AspNetCore.Mvc;
@@ -22,13 +23,19 @@ namespace api.Controllers
                 _context = context;
         }
         [HttpGet]
-        public async Task<IActionResult> GetAll(){
-            var stocks = await _stockRepo.GetAllAsync();
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            var stocks = await _stockRepo.GetAllAsync(query);
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stocks);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if(stock == null){
@@ -40,14 +47,20 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
             
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new {id = stockModel.Id}, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if(stockModel==null){
                 return NotFound();
@@ -57,8 +70,11 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stockModel = await _stockRepo.DeleteByIdAsync(id);
             if(stockModel == null){
                 return NotFound();
